bugfix / Add require for jQuery for Firefox compatibility
Add a require for version 1.7.1 of the minified jQuery. This is the
version of jQuery used on SE pages.

This script worked on Chrome prior to adding this require, but not on
Firefox. Now it works on Firefox too.

Also made a small change to the description to indicate that the script
is meant to be used on sites on the Stack Exchange network.