Use toggleClass() to control SBS toggling
Use jQuery's toggleClass() function to toggle the various CSS changes
for side-by-side mode.

This allows all of the CSS changes to be organized nicely near the top
of the script for easy changing.

Previously the script used a boolean dictionary to keep track of which
questions/answers were in side-by-side mode. This is no longer
necessary.

TODO: Fix "draft saved" bumping the side-by-side preview down on the Ask
Question page.