bugfix / new question preview pane
Fixes the behavior of the preview pane on the Ask Question page.

Before, the "draft saved" message would temporarily bump the preview
pane down while it was visible. Now we swap the order of these messages
and the preview pane to prevent this. Some extra css is added to help.

Also, changes the name of the script from "Side By Side Editing" to "SE
Side By Side Editing" to help with identification in the list of scripts
in greasemonkey/tampermonkey.